refactor purchase-filtering to use a custom source for generating data
